ADD show trigger
This makes it so that everytime the code triggers it gets updated in results.

This is a gimmecky way of doing this, a picture is downloaded and moved to the static folder. This would allow anyone to see the times the code triggered. In a future realease the storing of the triggers should be done in the database, with authentication.